feat: extract next runtime into separate binary
