+++ runtime/README.md
@@ -30,7 +30,7 @@ make test
 Load and run:
 
 ```bash
-cargo run -- --axum --provisioner-address http://localhost:5000
+cargo run --bin next -- --port 6001
 ```
 
 In another terminal:
@@ -40,24 +40,27 @@ In another terminal:
 grpcurl -plaintext -import-path ../proto -proto runtime.proto -d '{"service_name": "Tonic", "path": "/home/<path to shuttle>/runtime/axum.wasm"}' localhost:6001 runtime.Runtime/Load
 
 # start
-grpcurl -plaintext -import-path ../proto -proto runtime.proto -d '{"service_name": "Tonic", "deployment_id": "MDAwMDAwMDAtMDAwMC0wMDAwLTAwMDAtMDAwMDAwMDAwMDAw"}' localhost:6001 runtime.Runtime/Start
+grpcurl -plaintext -import-path ../proto -proto runtime.proto -d '{"deployment_id": "MDAwMDAwMDAtMDAwMC0wMDAwLTAwMDAtMDAwMDAwMDAwMDAw", "ip": "127.0.0.1:8000"}' localhost:6001 runtime.Runtime/Start
 
-# subscribe to logs (unimplemented)
+# subscribe to logs
 grpcurl -plaintext -import-path ../proto -proto runtime.proto localhost:6001 runtime.Runtime/SubscribeLogs
 
 # stop
-grpcurl -plaintext -import-path ../proto -proto runtime.proto -d '{"service_name": "Tonic", "deployment_id": "MDAwMDAwMDAtMDAwMC0wMDAwLTAwMDAtMDAwMDAwMDAwMDAw"}' localhost:6001 runtime.Runtime/Stop
+grpcurl -plaintext -import-path ../proto -proto runtime.proto -d '{}' localhost:6001 runtime.Runtime/Stop
 ```
 
 Curl the service:
 ```bash
-curl  localhost:7002/hello
+curl  localhost:8000/hello
 
-curl  localhost:7002/goodbye
+curl  localhost:8000/goodbye
 ```
 
 ## shuttle-legacy
-This will no loger load a `.so` will the code to start the runtime will be codegened for all services.
+
+This will no longer load a `.so` file, the code to start the runtime will be 
+codegened for all services.
+
 An example can be found in `src/bin/rocket.rs` which contains the secrets rocket example at the bottom and the codegen at the top.
 
 To test, first start a provisioner from the root directory using:
@@ -69,7 +72,7 @@ docker-compose -f docker-compose.rendered.yml up provisioner
 Then in another shell, start the wrapped runtime using the clap CLI:
 
 ```bash
-cargo run -- --port 6001 --storage-manager-type working-dir --storage-manager-path ./
+cargo run --bin rocket -- --port 6001 --storage-manager-type working-dir --storage-manager-path ./
 ```
 
 Or directly (this is the path hardcoded in `deployer::start`):
@@ -87,7 +90,7 @@ Then in another shell, load the service and start it up:
 grpcurl -plaintext -import-path ../proto -proto runtime.proto -d '{"service_name": "Tonic", "path": "/home/<path to shuttle>/examples/rocket/hello-world/target/debug/libhello_world.so", "secrets": {"MY_API_KEY": "test"}}' localhost:6001 runtime.Runtime/Load
 
 # run (this deployment id is default uuid encoded as base64)
-grpcurl -plaintext -import-path ../proto -proto runtime.proto -d '{"service_name": "Tonic", "deployment_id": "MDAwMDAwMDAtMDAwMC0wMDAwLTAwMDAtMDAwMDAwMDAwMDAw", "ip": "127.0.0.1:8000"}' localhost:6001 runtime.Runtime/Start
+grpcurl -plaintext -import-path ../proto -proto runtime.proto -d '{"deployment_id": "MDAwMDAwMDAtMDAwMC0wMDAwLTAwMDAtMDAwMDAwMDAwMDAw", "ip": "127.0.0.1:8000"}' localhost:6001 runtime.Runtime/Start
 
 # subscribe to logs
 grpcurl -plaintext -import-path ../proto -proto runtime.proto localhost:6001 runtime.Runtime/SubscribeLogs
