+++ runtime/src/lib.rs
@@ -1,9 +1,8 @@
-mod args;
-mod axum;
 mod legacy;
+mod next;
 mod provisioner_factory;
 
-pub use args::{Args, StorageManagerType};
-pub use axum::AxumWasm;
 pub use legacy::{start, Legacy};
+pub use next::AxumWasm;
+pub use next::NextArgs;
 pub use provisioner_factory::ProvisionerFactory;
